Added Stripe payment
Users will be charged when reservation requests are accepted by the parkingspot owner.
If the card is unable to be processed, the both parties will be notified that the card was declined,
and the request will not be approved or denied.

The transaction id as added to the reservation message, in case of cancellation at a later date for
refund.